Set up tests so that we can just test the CRUDServiceImpl instead of having to test each and every instance of that class for validation purposes.

+++ src/main/java/ca/corefacility/bioinformatics/irida/service/impl/CRUDServiceImpl.java
@@ -53,6 +53,10 @@ public class CRUDServiceImpl<KeyType, Type> implements CRUDService<KeyType, Type
 
     @Override
     public void delete(KeyType id) throws IllegalArgumentException {
+        if (!exists(id)) {
+            throw new IllegalArgumentException("No such identifier exists in the database.");
+        }
+
         repository.delete(id);
     }
